fabricdemo/api/pyproject.toml
[project]
name = "autonomous-network-demo-api"
version = "0.1.0"
description = "FastAPI backend for Autonomous Network NOC Demo"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.0.0",
    "azure-identity>=1.19.0",
    "azure-ai-projects>=1.0.0,<2.0.0",
    "azure-ai-agents==1.2.0b6",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",               # Required for FastAPI File/UploadFile
    "azure-storage-file-datalake>=12.14.0",  # OneLake CSV upload (Fabric Lakehouse)
    "azure-kusto-ingest>=4.3.0",             # Eventhouse KQL ingestion
]

[dependency-groups]
dev = []
